requirements.txt
playwright>=1.48.0
playwright-stealth>=1.0.6
pandas>=2.2.3
pyyaml>=6.0.1
beautifulsoup4>=4.12.3
requests>=2.31.0
rich>=13.7.0
tqdm>=4.66.1
phonenumbers>=8.13.27
